fix: preserve sparse monthly limit updates
